requirements.txt
asgiref==3.9.1
boto3==1.40.1
botocore==1.40.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2025.6
cssselect2==0.8.0
Django==5.2.4
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-storages==1.14.6
fonttools==4.59.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
psycopg2-binary==2.9.9
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
